refactor: 폴더 구조 1차 변경 (#52)
