+++ packages/components/Modal/ModalBody/ModalBody.styles.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const ModalBody = styled.section`
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+`;
