+++ packages/components/Modal/ModalFooter/ModalFooter.styles.ts
@@ -1,0 +1,8 @@
+import styled from '@emotion/styled';
+
+export const ModalFooter = styled.footer`
+  display: flex;
+  padding-top: 1rem;
+  justify-content: flex-end;
+  gap: 1rem;
+`;
